Asynchronize repository forking
