+++ src/main/scala/gitbucket/core/service/RepositoryCreationService.scala
@@ -6,8 +6,10 @@ import java.util.concurrent.ConcurrentHashMap
 import gitbucket.core.model.Profile.profile.blockingApi._
 import gitbucket.core.util.SyntaxSugars._
 import gitbucket.core.util.Directory._
-import gitbucket.core.util.JGitUtil
-import gitbucket.core.model.Account
+import gitbucket.core.util.{FileUtil, JGitUtil, LockUtil}
+import gitbucket.core.model.{Account, Role}
+import gitbucket.core.plugin.PluginRegistry
+import gitbucket.core.service.RepositoryService.RepositoryInfo
 import gitbucket.core.servlet.Database
 import org.apache.commons.io.FileUtils
 import org.eclipse.jgit.api.Git
@@ -126,15 +128,78 @@ trait RepositoryCreationService {
 
         // Record activity
         recordCreateRepositoryActivity(owner, name, loginUserName)
+
+        // Call hooks
+        PluginRegistry().getRepositoryHooks.foreach(_.created(owner, name))
       }
 
       RepositoryCreationService.endCreation(owner, name, None)
 
     } catch {
-      case ex: Exception => {
-        ex.printStackTrace()
-        RepositoryCreationService.endCreation(owner, name, Some(ex.toString))
+      case ex: Exception => RepositoryCreationService.endCreation(owner, name, Some(ex.toString))
+    }
+  }
+
+  def forkRepository(accountName: String, repository: RepositoryInfo, loginUserName: String): Future[Unit] = Future {
+    RepositoryCreationService.startCreation(accountName, repository.name)
+    try {
+      LockUtil.lock(s"${accountName}/${repository.name}") {
+        Database() withTransaction { implicit session =>
+          val originUserName = repository.repository.originUserName.getOrElse(repository.owner)
+          val originRepositoryName = repository.repository.originRepositoryName.getOrElse(repository.name)
+
+          insertRepository(
+            repositoryName = repository.name,
+            userName = accountName,
+            description = repository.repository.description,
+            isPrivate = repository.repository.isPrivate,
+            originRepositoryName = Some(originRepositoryName),
+            originUserName = Some(originUserName),
+            parentRepositoryName = Some(repository.name),
+            parentUserName = Some(repository.owner)
+          )
+
+          // Set default collaborators for the private fork
+          if (repository.repository.isPrivate) {
+            // Copy collaborators from the source repository
+            getCollaborators(repository.owner, repository.name).foreach { case (collaborator, _) =>
+              addCollaborator(accountName, repository.name, collaborator.collaboratorName, collaborator.role)
+            }
+            // Register an owner of the source repository as a collaborator
+            addCollaborator(accountName, repository.name, repository.owner, Role.ADMIN.name)
+          }
+
+          // Insert default labels
+          insertDefaultLabels(accountName, repository.name)
+          // Insert default priorities
+          insertDefaultPriorities(accountName, repository.name)
+
+          // clone repository actually
+          JGitUtil.cloneRepository(
+            getRepositoryDir(repository.owner, repository.name),
+            FileUtil.deleteIfExists(getRepositoryDir(accountName, repository.name)))
+
+          // Create Wiki repository
+          JGitUtil.cloneRepository(getWikiRepositoryDir(repository.owner, repository.name),
+            FileUtil.deleteIfExists(getWikiRepositoryDir(accountName, repository.name)))
+
+          // Copy LFS files
+          val lfsDir = getLfsDir(repository.owner, repository.name)
+          if (lfsDir.exists) {
+            FileUtils.copyDirectory(lfsDir, FileUtil.deleteIfExists(getLfsDir(accountName, repository.name)))
+          }
+
+          // Record activity
+          recordForkActivity(repository.owner, repository.name, loginUserName, accountName)
+
+          // Call hooks
+          PluginRegistry().getRepositoryHooks.foreach(_.forked(repository.owner, accountName, repository.name))
+
+          RepositoryCreationService.endCreation(accountName, repository.name, None)
+        }
       }
+    } catch {
+      case ex: Exception => RepositoryCreationService.endCreation(accountName, repository.name, Some(ex.toString))
     }
   }
